Adding Authorization Checks

+++ controllers/feed.js
@@ -134,6 +134,12 @@ exports.updatePost = (req, res, next) => {
                 throw error;
             }
 
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -162,6 +168,13 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+
             //Check logged user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
